requirements.txt
# This file is used to install the required packages for the project.flask
openai==0.28.0
langchain==0.3.21
langchain-core==0.3.45
langchain-community==0.3.20
langchain-text-splitters==0.3.7
chromadb==1.0.4
pydantic==2.7.4
tiktoken==0.9.0
GitPython==3.1.44
python-dotenv==1.1.0
Flask
